Successfully added nav base.
